Fix for SmartEnemy | Changes int to float to make it work

+++ src/com/tutorial/game/Game.java
@@ -100,7 +100,7 @@ public class Game extends Canvas implements Runnable {
         bs.show();
     }
 
-    public static int clamp(int var, int min, int max) {
+    public static float clamp(float var, float min, float max) {
         if (var >= max)
             return var = max;
         else if (var <= min)
