+++ src/com/tutorial/game/HUD.java
@@ -4,8 +4,8 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int HEALTH = 100;
-    private int colorTransitionValue = 255;
+    public static float HEALTH = 100;
+    private float colorTransitionValue = 255;
     private int score = 0;
     private int level = 1;
 
@@ -20,8 +20,8 @@ public class HUD {
     public void render(Graphics g) {
         g.setColor(Color.GRAY);
         g.fillRect(15, 15, 200, 32);
-        g.setColor(new Color(75, colorTransitionValue, 0));
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.setColor(new Color(75, (int)colorTransitionValue, 0));
+        g.fillRect(15, 15, (int)HEALTH * 2, 32);
         g.setColor(Color.WHITE);
         g.drawRect(15, 15, 200, 32);
